feat: add gen size comp
This was done by moving data from length comp to gen size comp. Note that
this did change the likelihood and mode results slightly, I think because
now the data is split into 2 different likelihood components, and so the
formulation of the likelihood fxn is slightly different. This means the
best solution is slightly different.